# internal/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: 社内OSSカタログ / 納品用 OSS 管理 API
  version: 1.0.0-draft2
  description: |
    社内で利用・納品対象となる OSS を一元管理し、プロジェクト毎の利用状況と納品用一覧（SPDX / CSV 等エクスポート）を生成するための API 初稿 (Phase 1)。

    **本稿の位置づけ**
    - Phase 1 対象: OSSコンポーネント/バージョン CRUD, タグ, プロジェクト, プロジェクト利用 (Usage), スコープ判定, ポリシー参照, 監査ログ最小, エクスポート(プレースホルダ)
    - 未実装(将来): 脆弱性, SBOM Import, NOTICE 生成, ライセンス全文管理, **JWT リフレッシュ**, 監査詳細検索高度化

    **表記**
    - ID は原則 UUID v4 文字列
    - 日本語説明は実装後英語化検討
servers:
  - url: https://api.example.local/api/v1
    description: 本番 (社内)
  - url: https://dev.api.example.local/api/v1
    description: 開発環境
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: 互換用途の Basic 認証（将来廃止予定）
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: アクセストークン用 JWT（リフレッシュなし）
  parameters:
    PageParam:
      name: page
      in: query
      description: 1 始まりのページ番号
      schema: { type: integer, minimum: 1, default: 1 }
    SizeParam:
      name: size
      in: query
      description: 1ページ件数 (最大 200)
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
  responses:
    NotFound:
      description: リソースが存在しない
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    Unauthorized:
      description: 認証失敗
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: 権限不足
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      description: RFC 9457 / RFC 7807 型エラー応答ボディ
      properties:
        type:
          type: string
          format: uri
          nullable: true
          description: 問題タイプ識別 URI（拡張分類用）
        title:
          type: string
          description: エラーの概要メッセージ（人間可読）
        status:
          type: integer
          description: HTTP ステータスコード
        detail:
          type: string
          nullable: true
          description: 追加詳細メッセージ
        instance:
          type: string
          format: uri
          nullable: true
          description: エラーが発生した具体的インスタンス URI（トレースID等）
        code:
          type: string
          nullable: true
          description: アプリケーション独自エラーコード
        errors:
          type: array
          description: フィールド単位バリデーションエラー配列
          items:
            type: object
            properties:
              field:
                type: string
                description: エラーが発生したフィールド名（JSON Pointer など）
              message:
                type: string
                description: フィールドに対するエラーメッセージ
      required: [title, status]

    # ---- JWT 関連追加 ----
...
    UsageRole:
      type: string
      enum:
        [
          BUNDLED_BINARY,
          BUNDLED_SOURCE,
          STATIC_LINK,
          DYNAMIC_LINK,
          RUNTIME_REQUIRED,
          SERVER_ENV,
          BUILD_ONLY,
          DEV_ONLY,
          TEST_ONLY,
        ]
      description: プロジェクト内での利用形態（配布対象か／工程限定か）
      x-enumDescriptions:
        - バイナリをそのまま成果物に同梱
        - ソースコードを成果物に同梱
        - 静的リンクにより最終バイナリへ取り込み
        - 動的ライブラリとして同梱/参照
        - 実行時環境として必要（顧客側準備含む）
        - サーバ運用環境のみで使用（配布物外）
        - ビルド工程のみで使用（成果物不含）
        - 開発専用（IDE/フォーマッタ等 配布物不含）
        - テスト工程専用（配布物不含）
    ScopeStatus:
      type: string
      enum: [IN_SCOPE, OUT_SCOPE, REVIEW_NEEDED]
      description: 納品対象スコープ判定状態（IN_SCOPE=含む, OUT_SCOPE=除外, REVIEW_NEEDED=要判定）
      x-enumDescriptions:
        - 納品対象に含める
        - 納品対象から除外
        - 判定保留（要レビュー）
...
    Project:
      type: object
      description: プロジェクト（納品単位）
      properties:
        id: { type: string, format: uuid, description: "プロジェクト ID" }
        projectCode: { type: string, description: "社内識別コード（ユニーク）" }
        name: { type: string, description: "プロジェクト名称" }
        department: { type: string, nullable: true, description: "担当部署" }
        manager:
          { type: string, nullable: true, description: "責任者 / 担当者名" }
        deliveryDate:
          {
            type: string,
            format: date,
            nullable: true,
            description: "納品予定日",
          }
        description:
          { type: string, nullable: true, description: "説明 / 備考" }
        ossUsageCount: { type: integer, description: "紐付く OSS 利用件数集計" }
        createdAt: { type: string, format: date-time, description: "作成日時" }
        updatedAt: { type: string, format: date-time, description: "更新日時" }
      required: [id, projectCode, name, createdAt, updatedAt]
    ProjectCreateRequest:
      type: object
      description: プロジェクト作成リクエスト
      required: [projectCode, name]
      properties:
        projectCode:
          { type: string, description: "ユニークなプロジェクトコード" }
        name: { type: string, description: "名称" }
        department: { type: string, nullable: true, description: "部署" }
        manager: { type: string, nullable: true, description: "責任者" }
        deliveryDate:
          { type: string, format: date, nullable: true, description: "納品日" }
        description:
          { type: string, nullable: true, description: "説明 / 備考" }
    ProjectUpdateRequest:
      type: object
      description: プロジェクト更新リクエスト
      properties:
        name: { type: string, description: "名称" }
        department: { type: string, nullable: true, description: "部署" }
        manager: { type: string, nullable: true, description: "責任者" }
        deliveryDate:
          { type: string, format: date, nullable: true, description: "納品日" }
        description:
          { type: string, nullable: true, description: "説明 / 備考" }
    ProjectUsage:
      type: object
      description: プロジェクトにおける特定 OSS バージョンの利用レコード
      properties:
        id: { type: string, format: uuid, description: "利用 ID" }
        projectId:
          { type: string, format: uuid, description: "プロジェクト ID" }
        ossId:
          {
            type: string,
            format: uuid,
            description: "OSSコンポーネント ID (利便性重複)",
          }
        ossVersionId:
          { type: string, format: uuid, description: "OSS バージョン ID" }
        usageRole:
          { $ref: "#/components/schemas/UsageRole", description: "利用形態" }
        scopeStatus:
          {
            $ref: "#/components/schemas/ScopeStatus",
            description: "スコープ判定",
          }
        inclusionNote:
          { type: string, nullable: true, description: "IN/OUT 判断理由メモ" }
        directDependency:
          { type: boolean, description: "直接依存なら true (間接は false)" }
        addedAt: { type: string, format: date-time, description: "登録日時" }
        evaluatedAt:
          {
            type: string,
            format: date-time,
            nullable: true,
            description: "スコープ判定更新日時",
          }
        evaluatedBy:
          { type: string, nullable: true, description: "判定実施ユーザ" }
      required:
        [
          id,
          projectId,
          ossId,
          ossVersionId,
          usageRole,
          scopeStatus,
          directDependency,
          addedAt,
        ]
    ProjectUsageCreateRequest:
      type: object
      description: プロジェクト利用作成リクエスト
      required: [ossId, ossVersionId, usageRole]
      properties:
        ossId:
          { type: string, format: uuid, description: "OSSコンポーネント ID" }
        ossVersionId:
          { type: string, format: uuid, description: "OSS バージョン ID" }
        usageRole:
          { $ref: "#/components/schemas/UsageRole", description: "利用形態" }
        directDependency:
          { type: boolean, default: true, description: "直接依存なら true" }
        inclusionNote:
          { type: string, nullable: true, description: "初期理由メモ" }
    ProjectUsageUpdateRequest:
      type: object
      description: プロジェクト利用更新リクエスト
      properties:
        ossVersionId:
          {
            type: string,
            format: uuid,
            description: "新しい OSS バージョン ID",
          }
        usageRole:
          { $ref: "#/components/schemas/UsageRole", description: "利用形態" }
        directDependency: { type: boolean, description: "直接依存フラグ" }
        inclusionNote: { type: string, nullable: true, description: "理由メモ" }
        scopeStatus:
          {
            $ref: "#/components/schemas/ScopeStatus",
            description: "スコープ判定 (手動上書き)",
          }
    ScopeStatusUpdateRequest:
      type: object
      description: スコープ判定更新リクエスト
      required: [scopeStatus]
      properties:
        scopeStatus:
          {
            $ref: "#/components/schemas/ScopeStatus",
            description: "更新後ステータス",
          }
        reasonCode:
          {
            type: string,
            nullable: true,
            description: "理由コード (マスタ化想定)",
          }
        reasonNote:
          { type: string, nullable: true, description: "自由記述理由" }
    ScopePolicy:
      type: object
      description: スコープ自動判定ポリシー設定
      properties:
        id: { type: string, format: uuid, description: "ポリシー ID" }
        runtimeRequiredDefaultInScope:
          {
            type: boolean,
            description: "RUNTIME_REQUIRED を自動 IN_SCOPE にするフラグ",
          }
        serverEnvIncluded:
          { type: boolean, description: "SERVER_ENV を IN_SCOPE とみなすか" }
        autoMarkForksInScope:
          {
            type: boolean,
            description: "INTERNAL_FORK を自動 IN_SCOPE 化するか",
          }
        updatedAt:
          { type: string, format: date-time, description: "最終更新日時" }
        updatedBy: { type: string, description: "更新ユーザ" }
    ScopePolicyUpdateRequest:
      type: object
      description: スコープポリシー更新リクエスト
      properties:
        runtimeRequiredDefaultInScope:
          {
            type: boolean,
            description: "RUNTIME_REQUIRED 自動 IN_SCOPE フラグ",
          }
        serverEnvIncluded:
          { type: boolean, description: "SERVER_ENV を IN_SCOPE とするか" }
        autoMarkForksInScope:
          { type: boolean, description: "INTERNAL_FORK 自動 IN_SCOPE フラグ" }
...
    PagedResult_Project:
      type: object
      description: プロジェクト一覧ページング結果
      properties:
        items:
          type: array
          description: 結果アイテム配列
          items: { $ref: "#/components/schemas/Project" }
        page: { type: integer, description: "現在ページ" }
        size: { type: integer, description: "ページサイズ" }
        total: { type: integer, description: "総件数" }
    PagedResult_ProjectUsage:
      type: object
      description: プロジェクト利用レコードのページング結果
      properties:
        items:
          type: array
          description: 結果アイテム配列
          items: { $ref: "#/components/schemas/ProjectUsage" }
        page: { type: integer, description: "現在ページ" }
        size: { type: integer, description: "ページサイズ" }
        total: { type: integer, description: "総件数" }

    # ---- USER / ROLE ----
...
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: プロジェクト一覧
      operationId: listProjects
      x-rolesAllowed: [VIEWER, EDITOR, ADMIN]
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
        - name: code
          in: query
          schema: { type: string }
        - name: name
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PagedResult_Project" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Projects]
      summary: プロジェクト作成
      operationId: createProject
      x-rolesAllowed: [EDITOR, ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectCreateRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /projects/{projectId}:
    get:
      tags: [Projects]
      summary: プロジェクト詳細
      operationId: getProject
      x-rolesAllowed: [VIEWER, EDITOR, ADMIN]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    patch:
      tags: [Projects]
      summary: プロジェクト更新
      operationId: updateProject
      x-rolesAllowed: [EDITOR, ADMIN]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectUpdateRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    delete:
      tags: [Projects]
      summary: プロジェクト削除 (論理予定)
      operationId: deleteProject
      x-rolesAllowed: [ADMIN]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: No Content }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /projects/{projectId}/usages:
    get:
      tags: [Project Usages]
      summary: プロジェクト中利用 OSS 一覧
      operationId: listProjectUsages
      x-rolesAllowed: [VIEWER, EDITOR, ADMIN]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
        - name: scopeStatus
          in: query
          schema: { $ref: "#/components/schemas/ScopeStatus" }
        - name: usageRole
          in: query
          schema: { $ref: "#/components/schemas/UsageRole" }
        - name: direct
          in: query
          description: 直接依存のみ true
          schema: { type: boolean }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PagedResult_ProjectUsage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Project Usages]
      summary: プロジェクト利用追加
      operationId: createProjectUsage
      x-rolesAllowed: [EDITOR, ADMIN]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectUsageCreateRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectUsage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /projects/{projectId}/usages/{usageId}:
    patch:
      tags: [Project Usages]
      summary: 利用情報更新
      operationId: updateProjectUsage
      x-rolesAllowed: [EDITOR, ADMIN]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: usageId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectUsageUpdateRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectUsage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    delete:
      tags: [Project Usages]
      summary: 利用削除
      operationId: deleteProjectUsage
      x-rolesAllowed: [EDITOR, ADMIN]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: usageId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: No Content }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /projects/{projectId}/usages/{usageId}/scope:
    patch:
      tags: [Project Usages]
      summary: スコープ判定更新
      operationId: updateProjectUsageScope
      x-rolesAllowed: [EDITOR, ADMIN]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: usageId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScopeStatusUpdateRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectUsage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /scope/policy:
    get:
      tags: [Scope Policy]
      summary: 現行スコープポリシー取得
      operationId: getScopePolicy
      x-rolesAllowed: [VIEWER, EDITOR, ADMIN]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScopePolicy" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    patch:
      tags: [Scope Policy]
      summary: スコープポリシー更新 (管理者)
      operationId: updateScopePolicy
      x-rolesAllowed: [ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScopePolicyUpdateRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScopePolicy" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /projects/{projectId}/export:
    get:
      tags: [Export]
      summary: プロジェクト納品用エクスポート (プレースホルダ)
      operationId: exportProjectArtifacts
      x-rolesAllowed: [VIEWER, EDITOR, ADMIN]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: format
          in: query
          required: true
          schema: { type: string, enum: [csv, spdx-json] }
        - name: scopes
          in: query
          schema: { type: string, description: "IN_SCOPE など (カンマ列挙)" }
      responses:
        "200":
          description: Export ファイル (現状 JSON/CSV テキスト)
          content:
            text/csv:
              schema: { type: string, format: binary }
            application/json:
              schema:
                type: object
                properties:
                  placeholder: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
# 追加予定 (将来)
# /import/sbom, /licenses, /notice, /vulnerabilities など